refactor: 파일 위치 이동 및 BaseTimeEntity 상속 코드 추가
