+++ src/main/java/com/example/tripKo/domain/address/entity/Address.java
@@ -1,6 +1,7 @@
 package com.example.tripKo.domain.address.entity;
 
 
+import com.example.tripKo.BaseTimeEntity;
 import lombok.*;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
@@ -12,7 +13,7 @@ import javax.persistence.*;
 @DynamicUpdate
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Address {
+public class Address extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false, updatable = false)
