+++ src/main/java/com/example/tripKo/domain/contents/entity/ContentsMenu.java
@@ -4,6 +4,7 @@ import static javax.persistence.FetchType.LAZY;
 import static javax.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
+import com.example.tripKo.BaseTimeEntity;
 import com.example.tripKo.domain.file.entity.File;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -23,7 +24,7 @@ import org.hibernate.annotations.DynamicUpdate;
 @Entity
 @NoArgsConstructor(access = PROTECTED)
 @Table(name = "contents_menu")
-public class ContentsMenu {
+public class ContentsMenu extends BaseTimeEntity {
 
   @Id
   @GeneratedValue(strategy = IDENTITY)
