+++ src/main/java/com/example/tripKo/domain/place/dao/PlaceJPARepository.java
@@ -1,5 +1,6 @@
-package com.example.tripKo.domain.place;
+package com.example.tripKo.domain.place.dao;
 
+import com.example.tripKo.domain.place.entity.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlaceJPARepository extends JpaRepository<Place, Integer> {
