+++ src/main/java/com/example/tripKo/domain/place/dao/PlaceRestaurantJPARepository.java
@@ -1,5 +1,6 @@
-package com.example.tripKo.domain.place.restaurant;
+package com.example.tripKo.domain.place.dao;
 
+import com.example.tripKo.domain.place.entity.PlaceRestaurant;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
