+++ src/main/java/com/example/tripKo/domain/place/dao/PlaceTouristSpotJPARepository.java
@@ -1,6 +1,6 @@
-package com.example.tripKo.domain.place.touristSpot;
+package com.example.tripKo.domain.place.dao;
 
-import com.example.tripKo.domain.place.festival.PlaceFestival;
+import com.example.tripKo.domain.place.entity.PlaceTouristSpot;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
