+++ src/main/java/com/example/tripKo/domain/place/dto/PlaceTouristSpotResponseDTO.java
@@ -1,5 +1,6 @@
-package com.example.tripKo.domain.place.touristSpot;
+package com.example.tripKo.domain.place.dto;
 
+import com.example.tripKo.domain.place.entity.PlaceTouristSpot;
 import lombok.Builder;
 import lombok.Getter;
